Convert storage to use fs.ErrNotExist and fs.PathErrors (#2531)
This converts the `storage` package to use `*fs.PathError` and
`fs.ErrNotExist` instead of a custom `storage.NewErrNotExist` and
`storage.IsNotExist`. This leans into the standard library, using the
same `errors` as other codebases. This makes it easier to express
not-exist errors in `Bucket` derivatives.

`*fs.PathError` is expected to have a value for `Op` - the stdlib
annoyingly does not handle an empty value for `Op` in its printout of
`Error()`:

```go
func (e *PathError) Error() string { return e.Op + " " + e.Path + ": " + e.Err.Error() }
```

We might want to file an issue for it to do so by not printing the
leading space. Regardless, for now, we lean into this and set the `Op`
to either `read` or `stat`, depending on what is most appropriate. This
is somewhat congruent with what the stdlib does itself. It seems that
the benefit of leaning into this stdlib error outweighs this slight
weirdness.

Q: Why didn't `storage` do this when it was written?
A: `storage` predates `io/fs`, so this didn't exist. Now that `io/fs` is
more of a standard (`os` uses it as well), this seems appropriate.