Change `buf alpha package` -> `buf alpha sdk` (#2535)
We've recently updated the language around this concept in the BSR -
update it on the CLI as well.

Because this is an alpha command, I didn't bother to forward the old
invocation to the new one, but I'm happy to add that. Also happy to add
a CHANGELOG entry, but it doesn't look like we've typically mentioned
changes to `alpha` commands in the past, with the exception of the
removal of `alpha registry token create`.

There's a few other spots where we mention remote packages in comments
or as field names in protos, but I don't think we want to include
updates to those in the scope of this work?