Clean up bufimage.ImageFile (#2550)
This PR does a few things:

- Removes `ImageFile.FileDescriptor`.
- Renames `ImageFile.Proto` to `ImageFile.FileDescriptorProto`.
- Moves `ImageFile.ImageFileWithIsImport` to a top-level function
`ImageFileWithIsImport`.
- Changes `protosource.InputFile.FileDescriptor` to
`FileDescriptorProto`.

This is a cleanup and simplification of the `ImageFile` interface in
preparation for our refactor of `bufmodule` and `bufimage`. `ImageFile`
had both `*descriptorpb.FileDescriptorProto` and
`protodescriptor.FileDescriptor` as properties on it, which was an
attempt to get out of the `FileDescriptorProto` struct business
entirely. This attempt failed. We still need
`protodescriptor.FileDescriptor` in some places, as it is the common
interface of `buf.alpha.image.v1.ImageFile` and
`google.protobuf.FileDescriptorProto`, and it also is preferable if it
can solely be used as it tries to show read-only by convention (even
though it isn't read-only, many of the fields returned are structs
themselves). However, with `ImageFile`, there's no need for a second
interface.