Update release workflow (#2552)
This updates the CLI release workflow. Releasing CLI itself (not
including downstream repos) now has three separate actions, each
requiring some level of manual approval.
1. Manually trigger the 'Create Release PR' action and the action will
create a PR that updates the changelog and version in `bufcli.go`.
2. Once the the first PR is merged, the second action is triggered. This
action builds and signs the new CLI and creates a draft release.
3. Once the draft release is published (approved), the third action is
triggered and creates a go-back-to-development PR that updates the
version and changelog.

This also removes third party actions in the release workflow.