Allow buf curl to use multiple schemas when resolving elements (#2587)
This enables the use of multiple `--schema` flags, and it also enables
using both `--reflect` and `--schema`. By default, if a `--schema` flag
is present, reflection is not used. But if one also explicitly specifies
`--reflect`, then both will be used.

This allows resolving elements in an RPC that may fall outside the
module that defines the RPC method. This can happen if responses
use extensions or google.protobuf.Any messages (which includes the
use of error details).