Prompt user to check for expired token in auth error message (#2630)
A user in our public slack got confused when they suddenly started
seeing this error message in the CLI:
> Failure: you are not authenticated. Create a new entry in your netrc,
using a Buf API Key as the password. For details, visit
https://docs.buf.build/bsr/authentication

The problem was that their token was expired and `buf registry logout`
fixed it for them.

This change improves the error message so that users know to check to
see if their token has expired.