Add buf migrate command (#2669)
This PR makes update to the migrate command:
* moves it from `buf beta migrate` to `buf migrate`
* Update flags to take `--module` and `--workspace` instead of paths to
those files.
* The destination for the buf.yaml and buf.lock is `.`, unless
`--workspace` is specified but `--module` is not specified, in which
case the destination is the directory for that `buf.yaml`.
- If the user runs `buf migrate --workspace foo/bar`, then the new
`buf.yaml` will be written at `foo/bar/buf.yaml`.
  - Otherwise, `./buf.yaml` will be written.
* Only allows one workspace to be specified. (This isn't new in this PR,
but want to highlight this).
* Add flag `--template`, allowing multiple paths to buf.gen.yamls.
* Fixes the bug where the command crashes when only one directory
without a buf.yaml is specified.
* Update lint/breaking rule translation to produce equivalent lists of
rules with minimal diff.

Note: with the recent change on [reading a lock file without
digest](https://github.com/bufbuild/buf/pull/2681). Right now `buf
migrate` fails with `Failure: unimplemented` if there is a buf.lock to
write. It shouldn't fail once commit service has been implemented.

---------

Co-authored-by: bufdev <bufdev-github@buf.build>