Re-introduce managed mode enabled warning (#2696)
We had a warning message where if the user has
managed mode options set but `enabled: false`: #1552.
This warning was originally implemented when reading
the config, which required the config reader to have access
to a logger.

Instead of passing around a logger to the config reader, we can
instead print this message when the `Generate` is called, since
the `generator` has access to a logger.