Fix `WorkspaceDepManager` to handle v1 module default behavior and rename `InputDirPath` to `SubDirPath` (#2802)
We were not defaulting to `v1` module behavior in `WorkspaceDepManager`
when we do not find a `v2` workspace before this. Now, if no `v2`
workspace
was found for `WorkspaceDepManager`, we treat the `subDirPath` as a `v1`
module
by default.
Also, this renamed `InputDirPath` to `SubDirPath` for clarity.
Lastly, this puts in a quick fix for the a `buf config init` test, since
the shape of the
default config has changed.