Improve bufmigrate error (#2821)
This returns the path of the failing config file that cannot be read
when
`buf config migrate` attempts to automatically gather workspace and
module configs by walking the input bucket. This gives the user clarity
on where the failing file is so they can either address it and/or set
flags
for `buf config migrate`.