Fix memory read bucket mapping when no controlling workspace is found (#2834)
In order to treat archive and git refs the same way we treat dir
and proto file refs, we are handling `inputSubDirPath` already
in workspace targeting. This means that we should not actually
be remapping archive and git memory buckets if no controlling
workspace is found.

This was found in our bug bash when testing against an archive
with a `v1` module rooted at the input sub dir path. Added a test
case for this.