Handle errors from new buf versions to old BSR versions
This adds a better error message / suggestions for remedying a situation
where it's suspected that a newer version of the CLI is interacting with
an older version of the BSR.

The set of cases to cover is:

1. Explicit code unimplemented - it's overwhelmingly likely that a newer
   CLI made a request to an unimplemented API in this case
2. Code Unknown + "Method Not Allowed" error message - this is an
   implementation detail of the BSR, but it's how older versions of the
   BSR will react to requests to new services
3. Code Unknown + http "Not Found", for new RPCs on known services. This
   is a result of the [default connect-go generated handlers][1].

This handles the first two cases. I'm not aware of any situations
currently where we hit 3, so have not added that here for lack of being
able to test (it would be trivial to extend the code unknown case if we
find one).

In newer versions of the BSR, we're going to look to change the behavior
resulting in 2, which will result in a separate case to handle here.

Here's a few examples of the output, running against an old version of
the BSR (specifically, v1.3.0).

Case 3:

```console
$ buf push
Failure: unknown: 405 Method Not Allowed for https://bufbuild.internal/buf.registry.module.v1.ModuleService/GetModules
This version of the buf CLI may require APIs that have not yet been deployed to https://bufbuild.internal
To resolve this failure, you can either:
- Try using an older version of the buf CLI
- Contact the site admin for https://bufbuild.internal to upgrade the instance
```

Case 1:

```console
$ buf dep update
Failure: unimplemented: 404 Not Found for https://bufbuild.internal/buf.registry.module.v1beta1.CommitService/GetCommits
This version of the buf CLI may require APIs that have not yet been deployed to https://bufbuild.internal
To resolve this failure, you can either:
- Try using an older version of the buf CLI
- Contact the site admin for https://bufbuild.internal to upgrade the instance
```

[1]: https://github.com/connectrpc/connect-go/blob/main/cmd/protoc-gen-connect-go/main.go