Provide mechanism to start implementing and testing editions-specific functionality (#2869)
This allows us to start implementing and testing logic for editions
(which is mostly confined to protosource and bufbreaking), before
actually having the CLI provide editions support to users. The idea is
that users won't be able to use editions with the CLI until everything
is updated to work correctly, but we need a way to enable the support
internally until then, for testing any in-progress editions-specific work.