Fix b5 digest calculation for remote modules
When a remote module is added to the workspace, the b5 digest of it
should not be calculated based on the state of other modules in the
workspace (which may point to later commits). Instead, the digest should
be used from the ModuleKey for the remote module (if the key references
a b5 digest) or it should be fetched from the CommitProvider.