Allow using WKT module as a dependency (#2906)
This fixes an issue where we were ignoring the use of
`buf.build/protocolbuffers/wellknowntypes`
as a configured dependency by always skipping WKT import paths and then
resolving the WKT
import at compile time.
This instead only skips the WKT path if it is not found in our module
set, and otherwise adds
the module containing the WKT path as a dependency.

Co-authored-by: bufdev <bufdev-github@buf.build>