Improve efficiency of resolver backed by protocompile output (#2915)
This changes the resolver provided from compilation to use the compiler's
symbol table as an index, to dead-reckon which file a symbol or extension
is in. This makes it drastically faster for modules with a very large number
of files. I benchmarked using the full googleapis (nearly 5000 files), and
for that case it's almost 100x faster.