Strip legacy options so they don't cause issues reparsing an image (#2913)
The Protobuf-Go runtime is unhappy if it is asked to create a
`protoreflect.FileDescriptor` for a file with any references to certain
legacy options: `MessageOptions.message_set_wire_format` and
`FieldOptions.weak`.

This updates `protoencoding.NewResolver` so that it can modify the
schema before creating the descriptors with the `protodesc` package. It
does **not** mutate the actual descriptor protos that are part of the
schema/image. So the schema is untouched, and the modifications are just
a way to force the runtime into allowing a resolver. The resolver is
typically used for looking up extensions/custom options and message
types in an Any, so these slight differences from the actual underlying
schema should not matter, except to 'buf convert' which has been updated
to disallow converting messages that contain message sets.