Update breaking checks for field types to consider message encoding feature (#2917)
This is for Editions support: with editions, just looking at the field
descriptor's `type` field won't tell you if the message is
length-prefixed (normal) or delimited (a la groups). But they are not
compatible. Usefully, the protobuf-go runtime implementation of
`FIeldDescriptor.Kind` _does_ return `GroupKind` when the
field uses delimited encoding, even though the underlying descriptor
proto would report "MESSAGE" as its type. So we leverage that, and
don't have to reinvent anything for resolving this feature.