Allow reconfiguration of extension and reserved ranges as long as no numbers are removed (#2916)
Previously, the breaking checks for extension and reserved ranges
required exactly the same ranges to be preserved. If a new schema
simply re-organized the ranges, but still covered the same span of
numbers, it would produce an error, even though it is technically
not a breaking change.

This PR introduces more analysis to identify the actual coverage
of the ranges. It is only an error if the new ranges, collectively, do
not cover all tag numbers that the previous ranges covered.