Add `--source-control-url` flag for buf push (#2930)
This PR adds the `--source-control-url` flag to `buf push`.
The `--source-control-url` flag allows the user to set a URL specific to
the contents
of the modules pushed (e.g. the specific commit to browse in source
control).

While the API allows for each module content to have a different source
control URL,
similar to labels, we are setting a single source control URL for all
module contents pushed.

---------

Co-authored-by: Philip K. Warren <pkwarren@users.noreply.github.com>