Add `--create-default-label` flag to `buf push` (#2931)
This adds the `--create-default-label` flag to `buf push`, which allows
the user to set a default label name for a module they are creating with
`--create`.
We do not allow this to be set if `--create` is not set, however, the
user does
not need to set this if they are using `--create`. If this is not set
during `--create`,
then the module will have a default label of "main".

---------

Co-authored-by: Philip K. Warren <pkwarren@users.noreply.github.com>