Fix location of map entries in bufprotosource (#2933)
This fixes the location reported for synthetic map entries and map entry
fields. They previously were reported without a location, which ends up
rendered as line 1, column 1 of the file. The problem is that the compiler
does not generate source location information for these synthetic
elements. So the bufprotosource elements (Message and Field) can
now look elsewhere if there is no source code location for the element
and the element is a synthetic map entry message or one of its fields.
The locations reported will now be the location of the corresponding
map field's type name. Unfortunately, that is as fine-grained as we can
get with what is actually in the source code info.