Add semantic checks related to syntax to "buf breaking" (#2934)
This adds four checks that all correspond with syntax, but will
correspond to "features" going forward, in Protobuf Editions.

1. The `MESSAGE_SAME_JSON_FORMAT` and `ENUM_SAME_JSON_FORMAT`
checks look at the `json_format` feature, which is currently only used
by the compiler. When it is set to "ALLOW" (for proto3 files and the default in
Editions), the compiler does more strict checks on the field names and
resulting computed JSON names to make sure there are no collisions. But
with the "LEGACY_BEST_EFFORT" value, most collisions are reported as
warnings and are not errors. This is the setting for proto2 files, since
there was no JSON format at the time proto2 was released. It can also be
configured this way in Editions via the feature (and is mainly intended
for use with migrating proto2 sources to Editions, and preserving their
semantics).
2. The `FIELD_SAME_UTF8_VALIDATION` check looks at the `utf8_validation`
feature. This captures the semantics where in proto2, string fields were
_supposed_ to be UTF8  but never actually verified in the runtime. In
proto3, on the other hand, the contents are verified. This can be configured
on a per-field level with Editions to allow migrating sources from proto2
(no validation) and then incrementally enabling validation.
3. The `ENUM_SAME_TYP` check looks at the `enum_type` feature, which
captures the difference between open (proto3 style) vs. closed (proto2 style)
enums. An open enum supports unknown numeric values. A
closed enum, on the other hand, does not. When an unknown value is
deserialized for a closed enum field, the value is stored in the
unrecognized fields set and the enum field is left unset/absent.

These checks are needed to identify backwards-incompatible changes to
features made in an Editions file. They can also capture incompatible
syntax changes, in the event that `FILE_SAME_SYNTAX` is configured to be
ignored (which should be acceptable to do if only using the Google-provided
runtimes/code generators).