Replace "label" check with "cardinality" check (#2927)
The name change is because some of the values of "cardinality" differ
from the label field in `FieldDescriptorProto`. Also, in "editions" some
of these values are configured via a `field_presence` feature, not the
field label.

The possible values for the field cardinality follow:
1. optional with implicit presence
2. optional with explicit presence
3. required
4. repeated
5. map

The old check is retained and marked deprecated, for backwards
compatibility. It is replaced by three new checks:
1. FIELD_SAME_CARDINALITY: for FILE and PACKAGE categories
2. FIELD_WIRE_COMPATIBLE_CARDINALITY: for WIRE caregory
3. FIELD_WIRE_JSON_COMPATIBLE_CARDINALITY: for WIRE_JSON category.

This also updates the oneof-related checks to ignore synthetic oneofs.
Synthetic oneofs are a side effect of proto3-optional fields. Changes to
them will now be covered by the cardinality check, which can now
distinguish between optional fields with implicit vs explicit field
presence (proto3-optional fields provide explicit presence in proto3).