Replace FIELD_SAME_CTYPE and FILE_SAME_JAVA_STRING_CHECK_UTF8 checks with Editions-aware versions (#2942)
The gist is that a new Java-specific utf8_validation feature has been
added to Editions that replaces the old java_string_check_utf8 file
option. Similarly, a new C++-specific string_type feature has been
added that replaces the old ctype field option. So the buf breaking
checks for those two options has now been updated to correctly
consider these features in Editions source files.

This one's not as simple as I had hoped. While some of the tricky bits
are in the protocompile repo's implementation of resolving features,
there's still non-trivial bits in here.

For one, the protobuf-go module's reflection support is quite
panic-happy. So in order to avoid panic issues, like getting an
extension value using a different descriptor than the one used to set
the value (even though they are for the same tag number and
describe the same value), this re-parses features in the options (by
serializing and then de-serializing), to make sure that the parsed
extension values of interest are associated with the "correct"
descriptor.

Also, the logic for computing whether a string is validated to be
UTF8 for Java is non-trivial. The logic isn't too many lines of code, but
it's subtle enough that this resulted in a very large set of files and
messages in the tests... so a large amount of test code for the new
FIELD_SAME_JAVA_UTF8_VALIDATION check.