Update protocompile to bring in a fix for buf breaking issue (#2968)
I was doing manual QA of buf breaking with an Editions file and ran into
something strange, related to the C++ string type and Java UTF8
validation checks.

The issue turned out to be a bug in the helper function in
`protocompile` that resolves custom features. This pulls in a fix to
that function.