Map target paths and target exclude paths for archive/git refs (#3003)
In CLI versions pre-1.32.0, for archive and git refs, we would map
the root of the bucket of archive and git refs to the `subDirPath`,
which is inconsistent with the behavior for dir refs (which maps
target paths and target exclude paths to the execution context).

In 1.32.0, we made changes to handle subDirPath, target paths,
and target exclude paths in a unified way, however this is not
backwards compatible with pre-1.32.0 versions. So, in order
to be backwards compatible, we need to remap the target
paths and target exclude paths.

I explored an option to check of the subDirPath contains
the target paths and target exclude paths before remapping,
however this is brittle, since we could hit an edge case with
directories that have the same name nested:

```
.
└── proto
    └── foo
        └── foo
```

Tests have been adjusted to reflect this change.

---------

Co-authored-by: Nick Snyder <nickdsnyder@gmail.com>
Co-authored-by: bufdev <4228796+bufdev@users.noreply.github.com>