Avoid building image for `buf export` when `--exclude-imports` flag is set (#3009)
This PR changes `buf export` so that when `--exclude-imports` is
passed, it does not require the input to be buildable as an image.
This allows a user to provide an input that may not be buildable, and
they can still get the source files back when `--exclude-imports` is
set, since we do not need the image to check for import files.

Fixes #3002