Warn when a plugin does not proto3 optional instead of error-ing (#3015)
This changes the behaviour of `buf generate` to warn when
a feature is not supported instead of erroring. And only error
in cases when edition is found.
This change is to keep backwards compatibility with pre-1.32.0
versions of the CLI, which would warn but ignore unsupported
features and continue to generate code.