Fix push `--git-metadata` flag for diverged checkouts (#3037)
This change switches back to `git ls-remote` to avoid using `git fetch
--tags`. Currently `git fetch --tags` will fail in environments where
the local tags point to different oids than the remote. This is
noticeable in CI systems where checkouts can reference the pseudoref of
an annotated tag causing fetch to fail. To help debug errors, from git
commands, we now also report the stderr output to the CLI error.