Include descriptor's location in set of ignore locations (#3051)
There are numerous places in the buf lint and breaking change rules
where a location is provided for some element within a descriptor --
like a name, number, option, etc. But the descriptor's `source_code_info`
only retains comments for complete declarations. So those other rules
are subject to the same potential issue, and the fix is to include the
descriptor's location as an ignore location in each of them.

Instead of updating every such rule, this change instead opts to address
it more systematically, by _always_ using the descriptor's location as
an ignore location (no need to explicitly pass it from every rule).