Make message literal output more consistent (#3031)
This canonicalizes the formatting of message literals in option values.
In particular, it will make the following changes to the format:
1. Always use `{` and `}`. If existing code uses `<` and `>`, change it.
2. Always include a separator between the field name and value. It is
technically not required between a field name and a message value or a
list value that contains messages. But it is not possible for the
formatter to know this perfectly for all cases (for example, in the case
of an empty list, how can the formatter know if its element type is a
message or not?). So, for consistency, this always uses a `:` separator.
If the existing code omits it, one is added.
3. Always omit optional separators between fields. The separators are
never necessary. Either `,` or `;` is accepted (you can even mix in the
same message literal, which is kinda confusing!). We were already
dropping superfluous separators if we ended up reformatting a message
into "compact" form (one-liner). Now we also do it for expanded form.