Only validate min/max editions in code gen response when input uses editions (#3073)
This works around an issue with some pre-v26.0 releases of protoc and
the code generators there in (such as Java, C++, Python, etc).

Some of the languages advertised support for features, but in an
experimental way, that did _not_ include the minimum and maximum edition
fields being set. With v1.32.0+ of buf, code gen with those plugins always
fails because the min/max editions are missing. This changes the validation
so that it only occurs when an actual input file uses editions. So if editions
aren't used at all, the plugin's output can be used without any error
messages.