Improve auth error for remote
Users interacting with a remote BSR may see an auth error when using the
env variable BUF_TOKEN. This is reported as:

Failure: the BUF_TOKEN environment variable is set, but is not valid.

To help users identify which remote was used, we now report the remote
in the error message:

Failure: the BUF_TOKEN environment variable is not valid for buf.build.

This can help clarify which remote was used. As an example this can
be caused when a modules dependencies are on a different BSR instance
then the parent module.