Remove username prompt on buf registry login (#3128)
This removes the requirement to provide your username when calling `buf
registry login`. On login the token is validated by fetching the current
user. The username is used only to validate that the current token is
associated with the user but is otherwise redundant. The flag has now
been marked deprecated and is hidden.

The output of the login command is also updated to be more user friendly
prompting the user to visit the web UI to create a token if they do not
have one. On successful login the user is informed that they are logged
in as the current user.