Remove deleteInvalidModuleData
This is an unnecesasry operation for the cache -- in the case where
there is invalid data, we would already be returning an error and the
contents of the cache will be flushed and reused. We also would need to
upgrade the lock from a shared lock to exclusive lock in that case.

This operation is still needed for the cache using tar files, so we
moved that logic up to getModuleDataForModuleKey.

This is likely also a cause for cache corruption, since it causes a
racey situation.