Validate domain in buf registry login
We don't do explicit validation of the domain in 'buf registry login',
which leads to some confusing error messages. Update the command to
validate the domain with netext.ValidateHostname and update
ValidateHostname to not accept URLs (it previously didn't look up the
port number so it didn't perform proper validation).