Add a comment with link to docs for `buf config init` (#3181)
This adds a comment at the top of the `buf.yaml` file generated from
`buf config init` with a link to our public docs to make it easier for
users
who are onboarding to get more information on `buf.yaml` configs.

Running `buf config init` generates a `buf.yaml` with the following
contents:

```
# For details on buf.yaml configuration, visit https://buf.build/docs/configuration/v2/buf-yaml
version: v2
lint:
  use:
    - DEFAULT
breaking:
  use:
    - FILE

```

The comment is preserved when round-tripping `buf.yaml`.

Fixes #3180