Omit stack trace from error level logs
As part of the migration to slog, stack traces are now computed and
included in logs at slog.LevelError and above. Update the zapslog
handler to pass slog.LevelError+1 as the level at which stack traces
should be included, which should better match the previous behavior of
buf when using zap.

In prior releases of buf before the migration to slog, zaputil.NewLogger
called zap.New without passing the zap.AddStacktrace option, which meant
that the addStackAt level defaulted to zapcore.FatalLevel + 1
(disabling the inclusion of stack traces in logs).