Fix Go feature set handling in editions for checks
This fixes the handling of the Go feature set in editions for use with
checks. Dynamic types would cause the google.golang.org/protobuf
library to panic when trying to fetch the Go feature extension. This is
a temporary workaround until upstream can be resolved. See the issue:
https://github.com/golang/protobuf/issues/1669

When resolving file descriptors for checks, we now check for use of
google/protobuf/go_features.proto and reparse extensions with a resolver
targetted at the gofeaturespb package. This ensure the types will match
the ones set in the extension type gofeaturespb.E_go.

Fixes #3580