Add input parameter filter for git commands
This adds a new input parameter filter to assist in working with large
git projects. The param filter sets the git filter-spec argument when
fetching repositories. Only one argument can be provided, the combined
form can be used to set multiple filters. When using filter combined
with a subdir the checkout will now also be sparse to further reduce the
number of workspace files.