Fix B5 digest calculations for transitive dependencies
This fixes the B5 digests calculation for remote Modules when a
transitive dependency is no longer a declared dependency in the target
ModuleSet. The digest calculation on Module.Digest would filter declared
dependencies of the remote Module to those of the parents. On creation
of bufmodule.NewModuleData the given getDeclaredDepModuleKeys must
return all direct and transistive dependencies. All declared
dependencies must be used when calculating the digest to ensure the
digests match for the Module and ModuleData.