Add exclude_options configuration for generation
This adds a new plugin option `exclude_options` for use on generation.
The excluded options are stripped from the Image creating a shallow
clone of the FileDescriptors that make up the ImageFiles. Any imports
that are no longer required due to options being removed are also
removed. The option is applied for the plugin. Multiple filtered images
that reference the original image may be created when exclude options
are set.