Experiment with finding the `--against` image using import paths
In order to figure out the corresponding `--against` image, this queries
the images for import paths within the set of `--against` images. This
provdies the following behaviours:

- If all files from the current image could not be found, then we simply
  ignore it.
- If files from the current image are found across multiple against
  images, we return an error.

This allows us to compare an input that may have a new module not
contained in the `--against` image set and ignore them, while still
using the proper per-module check configurations.