Add import of missing file on completion
If the completion item is not currently imported in the file, this will
add the missing import.

We can't necessarily assume that the file is already formatted by `buf
format`, so instead, we just either shove the import directly after any
existing imports (which will be reordered by `buf format`, if used); if
no imports, right after `package`; if no package, right after `syntax`,
and otherwise just balk.