LSP fix type completion on prefix modifiers
This improves the type completion to be more robust. Previously, we
would incorrectly ignore type completions when a tyep modifier was used
on the last line of a declaration. This was due to the declaration
identifying the field as the name field. Now, we split on the line with
both the type and name spans to avoid any parse issues.