Check for modules before attempting to create
This changes the logic to first check the modules before attempting to
create. This is because if a module already exists, we should not bother
to attempt to create it.

This also allows users that have pre-existing user-owned modules on
servers that now disallow creating new user-owned modules to continue
pushing to their modules.